requirements.txt
numpy==2.1.1
opencv-contrib-python==4.10.0.84
pillow==10.4.0
pygetwindow==0.0.9